refactor data
